+++ scripts/assets/playlist.js
@@ -5,9 +5,10 @@ import { CONSTANTS } from '../constants.js';
 /**
  * Extract assets from the given playlist
  * @param {Playlist|ClientDocumentMixin} playlist - The playlist to extract assets from.
+ * @param {Set<string>} selectedSounds - The sounds in the playlist.
  * @return {AssetData}
  */
-export async function ExtractPlaylistAssets(playlist) {
+export async function ExtractPlaylistAssets(playlist, selectedSounds) {
   const data = new AssetData({
     id: playlist?.id || '',
     name: playlist?.name || '',
@@ -18,11 +19,16 @@ export async function ExtractPlaylistAssets(playlist) {
     return data;
   }
 
-  const sounds = [];
+  let sounds = [];
 
   const playlistData = CONSTANTS.IsV10orNewer() ? playlist : playlist.data;
   if (playlistData.sounds?.size) {
     sounds.push(...Array.from(playlistData.sounds.values()));
+  }
+
+  if (selectedSounds.size) {
+    // Remove sounds that are not selected
+    sounds = sounds.filter(sound => selectedSounds.has(sound.id));
   }
 
   for (const sound of sounds) {